adding a command line option, -etd (enable threaded debugging), that uses a custom thread pool class to catch exceptions thrown inside of a thread.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@3450 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/utils/file/FSLock.java
@@ -35,7 +35,7 @@ public class FSLock {
      */
     public FSLock(File baseLocation) {
         lockFile = new File(baseLocation.getAbsoluteFile() + lockString);
-        lockFile.deleteOnExit();
+        if (lockFile != null) lockFile.deleteOnExit();
     }
 
     /**
